+++ api/data-source/events-data-source.ts
@@ -46,26 +46,7 @@ const events = [
   }
 ];
 
-const orders = [
-  {
-    id: uuid(),
-    total: 45,
-    shrimp_tortas_total: 1,
-    shrimp_torta_unitary_price: 25,
-    poc_chuc_tortas_total: 1,
-    poc_chuc_torta_unitary_price: 20,
-    event: {
-      id: "1",
-      created_at: 1000000000
-    },
-    owner: {
-      id: "1",
-      name: "Juan Perez"
-    },
-    paid: false,
-    canceled: false
-  }
-];
+const find = () => events;
 
 const addEvent = (event: any) => {
   event.id = uuid();
@@ -78,16 +59,4 @@ const updateEvent = (event: any, id: string, index: string | number) => {
   return (<any>events)[index];
 };
 
-const addOrder = (order: any) => {
-  order.id = uuid();
-  orders.push(order);
-  return order;
-};
-
-const updateOrder = (order: any, id: string, index: string | number) => {
-  (<any>orders)[index] = { id, ...order };
-  return (<any>orders)[index];
-};
-
-export const ordersDataSource = { orders, addOrder, updateOrder };
-export const eventsDataSource = { events, addEvent, updateEvent };
+export const eventsDataSource = { find, addEvent, updateEvent };
